+++ tests/formatter/input-output-pairs/test-corner-cases.in.gd
@@ -1,3 +1,12 @@
 class X:
 	func foo(result, end):  # TODO: reconsider
 		var path: String = result.path.substr(0, end).replace("res://", "").replace("tests/", "").replace("/", " ").capitalize()
+
+class Y extends X:
+	func bar():
+		pass
+
+class Z extends X:
+
+	func bar():
+		pass